Brain/_setup/mcp-server/pyproject.toml
[project]
name = "brain-mcp"
version = "0.1.0"
description = "MCP server exposing the Ai-Brain Obsidian vault as memory tools"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
brain-mcp = "brain_mcp.__main__:main"
brain-prep = "brain_mcp.brain_prep:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["brain_mcp*"]
